packages/evalgen/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-una>=0.5.1"]
build-backend = "hatchling.build"

[project]
name = "evalgen"
version = "0.0.1"
dynamic = ["una"]
requires-python = "~=3.13"
description = "Evalgen library for the Data Platform"
dependencies = [
    "databricks-sdk",
    "databricks-sqlalchemy>=2.0.6",
    "fastcore>=1.8.0",
    "fastsql>=2.0.3",
    "monsterui>=1.0.21",
    "polars>=1.12.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.19",
    "databricks-connect",
    "mlflow",
    "langchain>=0.3.25",
    "databricks-langchain>=0.4.0",
    "scikit-learn>=1.6.1",
    "databricks-agents>=0.17.2",
    "langchain-community>=0.3.21",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "basedpyright",
    "databricks-sdk",
    "databricks-connect",
    "pydantic",
    "ruff",
]
[tool.basedpyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.13"
[tool.uv.sources]

[tool.hatch.metadata]
allow-direct-references = true


[tool.hatch.build.hooks.una-build]
[tool.hatch.metadata.hooks.una-meta]
